Login and Loading a table now works.
when the user logs in the server will check if they have a table, if not then the blank state table becomes the table they work with. If the user disconnects, then the table will get saved as is, and next time they log in, it will get loaded from the saved file instead of starting from scratch.